+++ snapshot2/san-export.sh
@@ -6,7 +6,9 @@ set -uex
 ## rsync the files from raw-${SANTYPE} over to the current-${SANTYPE}
 ## snapshot current-${SANTYPE} to ${SANTYPE}-$DATE-ro
 ## delete the raw-${SANTYPE} snapshot
+JOB_NAME=${JOB_NAME:=db_backup}
 STORAGEEX="${TMPSTORAGE}/mysql"
+STORDUMP="${STOREAGE}/dumps/${SANTYPE}"
 [ -d ${STORAGEEX}/raw-${SANTYPE} ] &&  sudo btrfs sub delete ${STORAGEEX}/raw-${SANTYPE}
 sudo btrfs sub snapshot ${STORAGEEX}/current-raw ${STORAGEEX}/raw-${SANTYPE} && \
 sync && \
@@ -17,8 +19,15 @@ docker run -t --rm \
   ${DOCKERCON} \
   /media/infrastructure/snapshot2/san-export-con.sh ${SANTYPE} ${SANOUT}
 sync
-sudo rsync -avhP --delete ${STORAGEEX}/raw-${SANTYPE}/ ${STORAGE}/mysql/current-${SANTYPE}/
-[ ! -d ${STORAGE}/mysql/${SANTYPE}-$DATE-ro ] && \
-  sudo btrfs sub snapshot -r ${STORAGE}/mysql/current-${SANTYPE} ${STORAGE}/mysql/${SANTYPE}-${DATE}-ro
+
+## tar ball the *.txt files with the schema file
+if [ ${SANTYPE} != "redacted" ]; then
+  cd ${STORDUMP}/td && tar cjO - --use-compress-program=pbzip2 *.txt ${DBEXPORT}-tables.sql > ${STORDUMP}/${JOB_NAME}-${STAGE}.tar.bz2
+fi
+if [ ${SANTYPE} != "redacted" ]; then
+  sudo rsync -avhP --delete ${STORAGEEX}/raw-${SANTYPE}/ ${STORAGE}/mysql/current-${SANTYPE}/
+  [ ! -d ${STORAGE}/mysql/${SANTYPE}-$DATE-ro ] && \
+    sudo btrfs sub snapshot -r ${STORAGE}/mysql/current-${SANTYPE} ${STORAGE}/mysql/${SANTYPE}-${DATE}-ro
+fi
 sudo btrfs sub delete ${STORAGEEX}/raw-${SANTYPE}
 exit
